Use correct TLV content format number

+++ mbed-client/m2mbase.h
@@ -206,7 +206,7 @@ public:
      * \param content_type The content type to be set based on
      * CoAP specifications.
      */
-    virtual void set_coap_content_type(const uint8_t content_type);
+    virtual void set_coap_content_type(const uint16_t content_type);
 
     /**
      * \brief Sets the observable mode for the object.
@@ -326,7 +326,7 @@ public:
      * \brief Returns the CoAP content type of the object.
      * \return The CoAP content type of the object.
      */
-    virtual uint8_t coap_content_type() const;
+    virtual uint16_t coap_content_type() const;
 
     /**
      * \brief Returns the observation status of the object.
