+++ source/m2mnsdlinterface.cpp
@@ -1266,7 +1266,7 @@ void M2MNsdlInterface::send_resource_observation(M2MResource *resource,
         uint32_t token_length = 0;
 
         resource->get_observation_token(token,token_length);
-        uint8_t content_type = 0;
+        uint16_t content_type = 0;
         if(M2MResourceInstance::OPAQUE == resource->resource_instance_type()) {
             content_type = COAP_CONTENT_OMA_OPAQUE_TYPE;
         }
